+++ src/Validations/Contracts/CatalogImageValidationRules.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace PictaStudio\Contento\Validations\Contracts;
+
+interface CatalogImageValidationRules extends ProvidesValidationRules {}